requirements.txt
flask==3.0.3
Werkzeug>=3.0.0
tensorflow==2.15.0
pillow==10.4.0
numpy==1.26.4
scikit-image==0.24.0
opencv-python-headless==4.10.0.84
